+++ lib/internal/cluster/shared_handle.js
@@ -1,6 +1,6 @@
 'use strict';
 const assert = require('assert');
-const dgram = require('dgram');
+const dgram = require('internal/dgram');
 const net = require('net');
 
 module.exports = SharedHandle;
