+++ test/parallel/test-dgram-create-socket-handle.js
@@ -1,9 +1,9 @@
+// Flags: --expose-internals
 'use strict';
 const common = require('../common');
 const assert = require('assert');
-const dgram = require('dgram');
+const { _createSocketHandle } = require('internal/dgram');
 const UDP = process.binding('udp_wrap').UDP;
-const _createSocketHandle = dgram._createSocketHandle;
 
 // Throws if an "existing fd" is passed in.
 common.expectsError(() => {
